+++ spec/render/groovy_spec.rb
@@ -707,12 +707,32 @@ describe 'Render as Groovy' do
         ].join("\n")
       }
 
+      let(:actionwords_rendered) {
+        [
+          'package com.example',
+          '',
+          'class Actionwords extends ActionwordLibrary{',
+          '  def myProjectActionWord() {',
+          '  }',
+          '',
+          '  def myHighLevelProjectActionword() {',
+          '    myProjectActionWord()',
+          '  }',
+          '',
+          '  def myHighLevelActionword() {',
+          '    getDefaultLibrary().myFirstActionWord()',
+          '  }',
+          '}'
+        ].join("\n")
+      }
+
       let(:first_lib_rendered) {[
         'package com.example',
         '',
         '@Singleton',
         'class DefaultLibrary {',
         '  def myFirstActionWord() {',
+        '    // Tags: priority:high wip',
         '  }',
         '}'
       ].join("\n")}
@@ -723,6 +743,7 @@ describe 'Render as Groovy' do
         '@Singleton',
         'class WebLibrary {',
         '  def mySecondActionWord() {',
+        '    // Tags: priority:low done',
         '  }',
         '}'
       ].join("\n")}