+++ CHANGELOG.md
@@ -1,10 +1,26 @@
 Hiptest publisher changelog
 ===========================
 
+1.16.5
+------
+  - Fix UIDCall handling for behat
+
+1.16.4
+------
+  - Handling tags for shared actionwords
+
+1.16.3
+------
+  - Fix behat rendering of shared actionwords
+
+1.16.2
+------
+  - Remove unnecessary templates for cucumber/groovy
+
 1.16.0
 ------
   - Shared actionwords handling for behat
-  
+
 1.15.0
 ------
   - Shared actionwords handling for groovy/spock